changed amplitude to float

+++ earthquake_collector/models.py
@@ -8,7 +8,7 @@ class SeismicRecord(models.Model):
     longitude = models.DecimalField(max_digits=9, decimal_places=6)
 
     device_id = models.CharField(max_length=255)
-    amplitude = models.IntegerField(default=5)
+    amplitude = models.FloatField(default=5)
     added_at = models.DateTimeField(auto_now=True)
 
     def __str__(self):
